auto-annotate-cvat/requirements.txt
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
coloredlogs==15.0.1
contourpy==1.3.1
cvat-cli==2.24.0
cvat_sdk==2.24.0
cycler==0.12.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.12.0
future==1.0.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
Jinja2==3.1.5
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
sympy==1.13.3
tifffile==2024.12.12
tinydb==4.8.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
tuspy==0.2.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.53
ultralytics-thop==2.0.13
urllib3==2.3.0
